requirements.txt
accelerate==0.24.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blobfile==2.1.0
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
contourpy==1.1.1
coremltools==6.3.0
-e git+https://github.com/TomSchimansky/CustomTkinter.git@41f4aae01a54e7be569a313b5df5bbc561c597da#egg=customtkinter
cycler==0.12.1
Cython==3.0.4
darkdetect==0.3.1
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.21.4
docarray==0.39.1
docker==6.1.3
docker-pycreds==0.4.0
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.0
-e git+https://github.com/CASIA-IVA-Lab/FastSAM.git@38654b76b31d3178a8572579968f8eb404936318#egg=fastsam
ffmpy==0.3.1
filelock==3.13.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.40
googleapis-common-protos==1.61.0
gradio==3.35.2
gradio_client==0.6.1
grpcio==1.53.0
grpcio-health-checking==1.47.5
grpcio-reflection==1.47.5
grpcio-tools==1.53.0
guided-diffusion-sdk==1.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
invisible-watermark==0.2.0
jcloud==0.3
jina==3.22.3
jina-hubble-sdk==0.39.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kraken==4.3.13
lightning-utilities==0.9.0
linkify-it-py==2.0.2
llvmlite==0.39.1
lpips==0.1.4
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
Multi-Template-Matching==1.6.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
open-clip-torch==2.23.0
openai-clip==1.0.1
opencv-python==4.8.1.78
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-grpc==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-util-http==0.40b0
orjson==3.9.10
packaging==23.1
pandas==2.0.3
param==2.0.0
pathspec==0.11.2
pathtools==0.1.2
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
protobuf==4.21.12
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pycryptodomex==3.19.0
pydantic==1.10.13
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pyperclip==1.8.2
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-tkdnd==0.2.1
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
ratelim==0.1.6
referencing==0.30.2
regex==2023.10.3
replicate==0.15.4
requests==2.31.0
resize-right-sdk==1.0.2
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scipy==1.10.1
screeninfo==0.8.1
seaborn==0.13.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
-e git+https://github.com/Image-Py/sknw.git@68cf723834965217c40cf98ce82fe909a2213cfd#egg=sknw
smmap==5.0.1
sniffio==1.3.0
stability-sdk==0.8.4
stable-diffusion==0.0.0
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
timm==0.9.8
tk==0.1.0
tkintermapview==1.29
tokenizers==0.14.1
toolz==0.12.0
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers@9e87618f2be1401df55c36ad726629ae201e8e4d
trimesh==3.17.1
ttkwidgets==0.13.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.0.202
urllib3==1.26.18
uvicorn==0.23.1
wandb==0.15.12
watchfiles==0.21.0
wcwidth==0.2.8
websocket-client==1.6.1
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
